所有按鈕新增防呆機制

+++ src/composables/statCard.ts
@@ -122,9 +122,13 @@ export function useAuction() {
     console.log('點了' + item.treasureCode)
     addAttendance()
   }
-  const handleDeleteItem = (item: Treasure) => {
+  const handleDeleteItem = async (item: Treasure) => {
     submitDeleteTicketCode.value = item.treasureCode
-    deleteTreasure()
+    const result =  await useAlert.confirm("請確認是否要刪除此單?")
+    if (result.isConfirmed){
+      deleteTreasure()
+    }
+
   }
 
   const deleteTreasure = async () => {
@@ -256,7 +260,10 @@ export function useAuction() {
       return
     }
     if (!item.isBidding && item.canVerifyBiddingTicket){
-      confirmTicket()
+      const resulf = await useAlert.confirm("請確認是否收到帳款?")
+      if (resulf.isConfirmed){
+        confirmTicket()
+      }
       return
     }
     const result = await useAlert.confirm('你真的要標嗎???')
